// peso_lucky_admin-main/src/pages/Aviator/Order/Settled/data.d.ts
export interface Params {}
export interface ListItem {
	bet_amount: number;
	bet_direction: number;
	bet_odds: number;
	bet_type: number;
	created_at: number;
	currency: string;
	device_type: number;
	game_code: string;
	game_id: number;
	game_name: string;
	game_type: number;
	id: number;
	ip: string;
	merchant_code: string;
	merchant_id: number;
	merchant_name: string;
	merchant_user_id: string;
	nickname: string;
	order_number: string;
	prize_amount: number;
	seq: string;
	settle_time: number;
	site_id: number;
	status: number;
	updated_at: number;
	user_id: number;
	win_lose_amount: string;
}